tests/test_MochiKit-Selector.html

<html>
<head>
    <script type="text/javascript" src="../MochiKit/MockDOM.js"></script>
    <script type="text/javascript" src="../MochiKit/Base.js"></script>
    <script type="text/javascript" src="../MochiKit/Iter.js"></script>
    <script type="text/javascript" src="../MochiKit/DOM.js"></script>
    <script type="text/javascript" src="../MochiKit/Style.js"></script>    
    <script type="text/javascript" src="../MochiKit/Selector.js"></script>    
    <script type="text/javascript" src="SimpleTest/SimpleTest.js"></script>        
    <link rel="stylesheet" type="text/css" href="SimpleTest/test.css">
    <style type="text/css">
        p {
            display: none;
        } 
    </style>
</head>
<body>
    <p>Test originally from <a href="http://simon.incutio.com/archive/2003/03/25/#getElementsBySelector" rel="bookmark">this blog entry</a>.</p>

    <p>Here are some links in a normal paragraph: <a href="http://www.google.com/" title="Google!">Google</a>, <a href="http://groups.google.com/">Google Groups</a>. This link has <code>class="blog"</code>: <a href="http://diveintomark.org/" class="blog">diveintomark</a></p>
    <div id="foo">
    <p>Everything inside the red border is inside a div with <code>id="foo"</code>.</p>
    <p>This is a normal link: <a href="http://www.yahoo.com/">Yahoo</a></p>
    
    <a style="display: none" href="http://www.example.com/outsidep">This a is not inside a p</a>

    <p>This link has <code>class="blog"</code>: <a href="http://simon.incutio.com/" class="blog">Simon Willison's Weblog</a></p>
    <p>This <span><a href="http://www.example.com/insidespan">link</a></span> is inside a span, not directly child of p</p>
    <p lang="en-us">Nonninn</p>
    <p lang="is-IS">Sniðugt</p>
    </div>
<pre id="test">
<script type="text/javascript">
try {

    var testExpected = function (res, exp, lbl) {
        for (var i=0; i < res.length; i ++) {
            is( res[i].href, exp[i], lbl + ' (' + i + ')');
        }
    };

    var expected = ['http://simon.incutio.com/archive/2003/03/25/#getElementsBySelector',
                    'http://www.google.com/',
                    'http://groups.google.com/',
                    'http://diveintomark.org/',
                    'http://www.yahoo.com/',
                    'http://www.example.com/outsidep',
                    'http://simon.incutio.com/',
                    'http://www.example.com/insidespan'];
    var results = $$('a');
    testExpected(results, expected, "'a' selector");

    expected = ['http://diveintomark.org/', 'http://simon.incutio.com/'];
    results = $$('p a.blog');    
    testExpected(results, expected, "'p a.blog' selector");

    expected = ['http://www.yahoo.com/',
                'http://www.example.com/outsidep',
                'http://simon.incutio.com/',
                'http://www.example.com/insidespan'];
    results = $$('div a');
    testExpected(results, expected, "'div a' selector");

    expected = ['http://www.yahoo.com/',
                'http://www.example.com/outsidep',
                'http://simon.incutio.com/',
                'http://www.example.com/insidespan'];
    results = $$('div#foo a');
    testExpected(results, expected, "'div#foo a' selector");

    expected = ['http://simon.incutio.com/',
                'http://www.example.com/insidespan'];
    results = $$('#foo a.blog');
    testExpected(results, expected, "'#foo a.blog' selector");

    expected = ['http://diveintomark.org/',
                'http://simon.incutio.com/',
                'http://www.example.com/insidespan'];
    results = $$('.blog');
    testExpected(results, expected, "'.blog' selector");

    expected = ['http://www.google.com/',
                'http://www.yahoo.com/',
                'http://www.example.com/outsidep',
                'http://www.example.com/insidespan'];
    results = $$('a[href^="http://www"]');
    testExpected(results, expected, "'a[href^=http://www]' selector");

    expected = ['http://diveintomark.org/'];
    results = $$('a[href$="org/"]');
    testExpected(results, expected, "'a[href$=org/]' selector");

    expected = ['http://www.google.com/',
                'http://groups.google.com/'];
    results = $$('a[href*="google"]');
    testExpected(results, expected, "'a[href*=google]' selector");

    expected = ['http://simon.incutio.com/archive/2003/03/25/#getElementsBySelector'];
    results = $$('a[rel="bookmark"]');
    testExpected(results, expected, "'a[rel=bookmark]' selector");

    expected = ['http://www.google.com/'];
    results = $$('a[title]');
    testExpected(results, expected, "'a[title]' selector");

    results = $$('p[lang|="en"]');
    is( results[0].firstChild.nodeValue, 'Nonninn', "'p[lang|=en]' selector");

    expected = ['http://simon.incutio.com/archive/2003/03/25/#getElementsBySelector',
                'http://www.google.com/',
                'http://groups.google.com/',
                'http://diveintomark.org/',
                'http://www.yahoo.com/',
                'http://simon.incutio.com/',
                'http://www.example.com/insidespan'];
    results = $$('p > a');
    testExpected(results, expected, "'p > a' selector");

    expected = ['http://www.example.com/insidespan'];
    results = $$('span > a');
    testExpected(results, expected, "'span > a' selector");

    expected = ['http://groups.google.com/'];
    results = $$('a + a');
    testExpected(results, expected, "'a + a' selector");

    expected = document.getElementsByTagName('p');
    results = $$('a[href$=outsidep] ~ *');
    for (var i=0; i < results.length; i ++) {
        is( results[i], expected[i+4], "'a[href$=outsidep] ~ *' selector" + ' (' + i + ')');
    }

    ok( true, "test suite finished!");
    
    
} catch (err) {
    
    var s = "test suite failure!\n";
    var o = {};
    var k = null;
    for (k in err) {
        // ensure unique keys?!
        if (!o[k]) {
            s +=  k + ": " + err[k] + "\n";
            o[k] = err[k];
        }
    }
    ok ( false, s );

}
</script>
</pre>
</body>
</html>
